requirements.txt
geopandas>=0.14
folium>=0.15
pandas>=2.0
numpy>=1.24
rasterio>=1.3
shapely>=2.0
matplotlib>=3.8
geemap>=0.30
earthengine-api>=0.1.400
jinja2>=3.1
python-dotenv>=1.0
